requirements.txt
cffi==1.8.3
click==6.6
cryptography==1.5.2
Flask==0.11.1
idna==2.1
itsdangerous==0.24
Jinja2==2.8
labsuite==0.4
MarkupSafe==0.23
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.1.0
PyYAML==3.12
six==1.10.0
Werkzeug==0.11.11
